Moved the request classes with the cached body and input stream.
